TGGeometry

Documentation for TGGeometry.

TGGeometry.TGGeom

TGGeometry.tg_geom_type

TGGeometry.tg_index

TGGeometry.tg_point

TGGeometry.tg_rect

TGGeometry.tg_segment

TGGeometry.contains

TGGeometry.coveredby

TGGeometry.covers

TGGeometry.disjoint

TGGeometry.equals

TGGeometry.intersects

TGGeometry.tg_env_set_allocator

TGGeometry.tg_geom_equals

TGGeometry.tg_geom_geojson

TGGeometry.tg_geom_new_point

TGGeometry.tg_geom_new_point_z

TGGeometry.tg_geom_typeof

TGGeometry.tg_line_new

TGGeometry.tg_parse_geojson

TGGeometry.tg_point_rect

TGGeometry.tg_poly_new

TGGeometry.tg_rect_expand

TGGeometry.tg_ring_new

TGGeometry.tg_segment_rect

TGGeometry.touches

TGGeometry.within

TGGeometry.TGGeom — Type

TGGeom

A wrapper around a tg_geom pointer. Automatically handles memory management through a finalizer.

TGGeometry.tg_geom_type — Type

tg_geom_type

Geometry types.

All tg_geom are one of the following underlying types.

Enumerator	Note

TG_POINT	Point

TG_LINESTRING	LineString

TG_POLYGON	Polygon

TG_MULTIPOINT	MultiPoint, collection of points

TG_MULTILINESTRING	MultiLineString, collection of linestrings

TG_MULTIPOLYGON	MultiPolygon, collection of polygons

TG_GEOMETRYCOLLECTION	GeometryCollection, collection of geometries

See also

tg_geom_typeof(), tg_geom_type_string(), GeometryAccessors

TGGeometry.tg_index — Type

tg_index

Geometry indexing options.

Used for polygons, rings, and lines to make the point-in-polygon and geometry intersection operations fast.

An index can also be used for efficiently traversing, searching, and performing nearest-neighbor (kNN) queries on the segment using tg_ring_index_*() and tg_ring_nearest() functions.

Enumerator	Note

TG_DEFAULT	default is TG_NATURAL or tg_env_set_default_index().

TG_NONE	no indexing available, or disabled.

TG_NATURAL	indexing with natural ring order, for rings/lines

TG_YSTRIPES	indexing using segment striping, rings only

TGGeometry.tg_point — Type

tg_point

The base point type used for all geometries.

See also

PointFuncs

TGGeometry.tg_rect — Type

tg_rect

A rectangle defined by a minimum and maximum coordinates. Returned by the tg_geom_rect(), tg_ring_rect(), and other *_rect() functions for getting a geometry's minumum bounding rectangle. Also used internally for geometry indexing.

See also

RectFuncs

TGGeometry.tg_segment — Type

tg_segment

The base segment type used in tg_line and tg_ring for joining two vertices.

See also

SegmentFuncs

TGGeometry.contains — Function

contains(geom1, geom2)

Check if geom1 completely contains geom2, as defined by DE-9IM [de9im].

This means that the boundary and interior of geom2 lie entirely inside geom1, and they cannot intersect the exterior of geom1.

Arguments

geom1: The first geometry. May be any GeoInterface-compatible geometry.

geom2: The second geometry. May be any GeoInterface-compatible geometry.

TGGeometry.coveredby — Function

coveredby(geom1, geom2)

Check if geom1 is covered by geom2, as defined by DE-9IM [de9im].

This means that no points of geom1 lie outside of geom2.

Equivalent to covers(geom2, geom1) - argument order reversed.

Arguments

geom1: The first geometry. May be any GeoInterface-compatible geometry.

geom2: The second geometry. May be any GeoInterface-compatible geometry.

TGGeometry.covers — Function

covers(geom1, geom2)

Check if geom1 covers geom2, as defined by DE-9IM [de9im].

This means that no points of geom2 lie outside of geom1.

Arguments

geom1: The first geometry. May be any GeoInterface-compatible geometry.

geom2: The second geometry. May be any GeoInterface-compatible geometry.

TGGeometry.disjoint — Function

disjoint(geom1, geom2)

Check if geom1 and geom2 are disjoint (completely non-intersecting and separate), as defined by DE-9IM [de9im].

Arguments

geom1: The first geometry. May be any GeoInterface-compatible geometry.

geom2: The second geometry. May be any GeoInterface-compatible geometry.

TGGeometry.equals — Function

equals(geom1, geom2)

Check if geom1 and geom2 are spatially equal, as defined by DE-9IM [de9im].

Arguments

geom1: The first geometry. May be any GeoInterface-compatible geometry.

geom2: The second geometry. May be any GeoInterface-compatible geometry.

TGGeometry.intersects — Function

intersects(geom1, geom2)

Check if geom1 and geom2 intersect, as defined by DE-9IM [de9im].

Specifically, this checks that the intersection of two geometries does not result in an empty set.

Arguments

geom1: The first geometry. May be any GeoInterface-compatible geometry.

geom2: The second geometry. May be any GeoInterface-compatible geometry.

TGGeometry.tg_env_set_allocator — Method

tg_env_set_allocator(malloc, realloc, free)

GlobalFuncs Global environment

Functions for optionally setting the behavior of the TG environment. These, if desired, should be called only once at program start up and prior to calling any other tg_*() functions. @{

TGGeometry.tg_geom_equals — Method

tg_geom_equals(a, b)

GeometryPredicates Geometry predicates

Functions for testing the spatial relations of two geometries. @{

TGGeometry.tg_geom_geojson — Method

tg_geom_geojson(geom, dst, n)

GeometryWriting Geometry writing

Functions for writing geometries as external data representations. @{

TGGeometry.tg_geom_new_point — Method

tg_geom_new_point(point)

GeometryConstructors Geometry constructors

Functions for creating and freeing geometries. @{

TGGeometry.tg_geom_new_point_z — Method

tg_geom_new_point_z(point, z)

GeometryConstructorsEx Geometry with alternative dimensions

Functions for working with geometries that have more than two dimensions or are empty. The extra dimensional coordinates contained within these geometries are only carried along and serve no other purpose than to be available for when it's desired to export to an output representation such as GeoJSON, WKT, or WKB. @{

TGGeometry.tg_geom_typeof — Method

tg_geom_typeof(geom)

GeometryAccessors Geometry accessors

Functions for accessing various information about geometries, such as getting the geometry type or extracting underlying components or coordinates. @{

TGGeometry.tg_line_new — Method

tg_line_new(points, npoints)

LineFuncs Line functions

Functions for working directly with the tg_line type.

There are no direct spatial predicates for tg_line. If you want to perform operations like "intersects" or "covers" then you must upcast the line to a tg_geom, like such:

[`tg_geom_intersects`](@ref)((struct [`tg_geom`](@ref)*)line, geom); @{

TGGeometry.tg_parse_geojson — Method

tg_parse_geojson(geojson)

GeometryParsing Geometry parsing

Functions for parsing geometries from external data representations. It's recommended to use tg_geom_error() after parsing to check for errors. @{

TGGeometry.tg_point_rect — Method

tg_point_rect(point)

PointFuncs Point functions

Functions for working directly with the tg_point type. @{

TGGeometry.tg_poly_new — Method

tg_poly_new(exterior, holes, nholes)

PolyFuncs Polygon functions

Functions for working directly with the tg_poly type.

There are no direct spatial predicates for tg_poly. If you want to perform operations like "intersects" or "covers" then you must upcast the poly to a tg_geom, like such:

[`tg_geom_intersects`](@ref)((struct [`tg_geom`](@ref)*)poly, geom); @{

TGGeometry.tg_rect_expand — Method

tg_rect_expand(rect, other)

RectFuncs Rectangle functions

Functions for working directly with the tg_rect type. @{

TGGeometry.tg_ring_new — Method

tg_ring_new(points, npoints)

RingFuncs Ring functions

Functions for working directly with the tg_ring type.

There are no direct spatial predicates for tg_ring. If you want to perform operations like "intersects" or "covers" then you must upcast the ring to a tg_geom, like such:

[`tg_geom_intersects`](@ref)((struct [`tg_geom`](@ref)*)ring, geom); @{

TGGeometry.tg_segment_rect — Method

tg_segment_rect(s)

SegmentFuncs Segment functions

Functions for working directly with the tg_segment type. @{

TGGeometry.touches — Function

touches(geom1, geom2)

Check if geom1 touches geom2, as defined by DE-9IM [de9im].

This means that the intersection of two geometries has at least one point in common, but does not completely contain the other.

Arguments

geom1: The first geometry. May be any GeoInterface-compatible geometry.

geom2: The second geometry. May be any GeoInterface-compatible geometry.

TGGeometry.within — Function

within(geom1, geom2)

Check if geom1 is within geom2, as defined by DE-9IM [de9im].

This means that the boundary and interior of geom1 lie entirely inside geom2, and they cannot intersect the exterior of geom2.

This is the opposite of contains(geom2, geom1).

Arguments

geom1: The first geometry. May be any GeoInterface-compatible geometry.

geom2: The second geometry. May be any GeoInterface-compatible geometry.

de9imthe Dimensionally Extended 9-Intersection Model that describes the relationship between two planar geometries.

de9imthe Dimensionally Extended 9-Intersection Model that describes the relationship between two planar geometries.

de9imthe Dimensionally Extended 9-Intersection Model that describes the relationship between two planar geometries.

de9imthe Dimensionally Extended 9-Intersection Model that describes the relationship between two planar geometries.

de9imthe Dimensionally Extended 9-Intersection Model that describes the relationship between two planar geometries.

de9imthe Dimensionally Extended 9-Intersection Model that describes the relationship between two planar geometries.

de9imthe Dimensionally Extended 9-Intersection Model that describes the relationship between two planar geometries.

de9imthe Dimensionally Extended 9-Intersection Model that describes the relationship between two planar geometries.